cheking

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -11,21 +11,21 @@
 int main(void)
 {
 int y;
-long int u = 1;
-long int p = 2;
-long int k;
-printf("%ld, ", u);
-printf("%ld, ", p);
+unsigned int u = 1;
+unsigned int p = 2;
+unsigned int k;
+printf("%u, ", u);
+printf("%u, ", p);
 for (y = 1 ; y < 97 ; y++)
 {
 k = p + u;
 if (y != 96)
 {
-printf("%ld, ", k);
+printf("%u, ", k);
 }
 else
 {
-printf("%ld", k);
+printf("%u", k);
 }
 u = p;
 p = k;
